Add files for lab 6
Code files for visualizing the waypoints in rviz are added.

These waypoints are generated from running the car through the Levine loop
once, and running the `waypoint_logger.py` node. Since there are A LOT
of waypoints generated and computing power is limited, the full .csv
data is filtered to only contain important points. This filtering is
done by a Jupyter notebook.